requirements.txt
streamlit
pandas
yfinance
google-generativeai